testes realizados

+++ src/CapitalGain.Domain/Entities/Money.cs
@@ -8,7 +8,7 @@ public class Money
     [JsonPropertyName("operation")]
     public string Operation { get; set; }="";
     [JsonPropertyName("unit-cost")]
-    public double UnitCost { get; set; }
+    public decimal UnitCost { get; set; }
     [JsonPropertyName("quantity")]
     public int Quantity { get; set; }
